feat: new folder structure for splitter
