+++ src/components/d-split-screen/splitter/DSplitter.style.scss
@@ -1,5 +1,5 @@
-@import "../../styles/helpers";
-@import "../../styles/media";
+@import "../../../styles/helpers";
+@import "../../../styles/media";
 
 .d-splitter {
   position: absolute;
